--- server/schedule/checker/split_merge_test.go
@@ -1,94 +1,0 @@
-// Copyright 2019 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package checker
-
-import (
-	"testing"
-	"time"
-
-	. "github.com/pingcap/check"
-	"github.com/pingcap/kvproto/pkg/metapb"
-	"github.com/pingcap/pd/pkg/mock/mockcluster"
-	"github.com/pingcap/pd/pkg/mock/mockoption"
-	"github.com/pingcap/pd/server/core"
-	"github.com/pingcap/pd/server/namespace"
-)
-
-func TestSplitMerge(t *testing.T) {
-	TestingT(t)
-}
-
-var _ = Suite(&testSplitMergeSuite{})
-
-type testSplitMergeSuite struct{}
-
-func (s *testMergeCheckerSuite) TestCache(c *C) {
-	cfg := mockoption.NewScheduleOptions()
-	cfg.MaxMergeRegionSize = 2
-	cfg.MaxMergeRegionKeys = 2
-	cfg.SplitMergeInterval = time.Hour
-	s.cluster = mockcluster.NewCluster(cfg)
-	stores := map[uint64][]string{
-		1: {}, 2: {}, 3: {}, 4: {}, 5: {}, 6: {},
-	}
-	for storeID, labels := range stores {
-		s.cluster.PutStoreWithLabels(storeID, labels...)
-	}
-	s.regions = []*core.RegionInfo{
-		core.NewRegionInfo(
-			&metapb.Region{
-				Id:       2,
-				StartKey: []byte("a"),
-				EndKey:   []byte("t"),
-				Peers: []*metapb.Peer{
-					{Id: 103, StoreId: 1},
-					{Id: 104, StoreId: 4},
-					{Id: 105, StoreId: 5},
-				},
-			},
-			&metapb.Peer{Id: 104, StoreId: 4},
-			core.SetApproximateSize(200),
-			core.SetApproximateKeys(200),
-		),
-		core.NewRegionInfo(
-			&metapb.Region{
-				Id:       3,
-				StartKey: []byte("t"),
-				EndKey:   []byte("x"),
-				Peers: []*metapb.Peer{
-					{Id: 106, StoreId: 2},
-					{Id: 107, StoreId: 5},
-					{Id: 108, StoreId: 6},
-				},
-			},
-			&metapb.Peer{Id: 108, StoreId: 6},
-			core.SetApproximateSize(1),
-			core.SetApproximateKeys(1),
-		),
-	}
-
-	for _, region := range s.regions {
-		s.cluster.PutRegion(region)
-	}
-
-	s.mc = NewMergeChecker(s.cluster, namespace.DefaultClassifier)
-
-	ops := s.mc.Check(s.regions[1])
-	c.Assert(ops, IsNil)
-	s.cluster.SplitMergeInterval = 0
-	time.Sleep(time.Second)
-	ops = s.mc.Check(s.regions[1])
-	c.Assert(ops, NotNil)
-
-}
